% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_nb.R
\name{calc_pi_nb_analytic}
\alias{calc_pi_nb_analytic}
\title{Estimate glm.nb prediction interval analytically}
\usage{
calc_pi_nb_analytic(
  df,
  fit,
  quantile_levels = c(0.01, 0.025, seq(0.05, 0.95, by = 0.05), 0.975, 0.99)
)
}
\arguments{
\item{df}{dataframe to get predictions on}

\item{fit}{model fit via MASS::glm.nb}

\item{quantile_levels}{quantiles that you want prediction intervals at}
}
\value{
matrix of prediction intervals for each row of df
}
\description{
Internal function to get prediction intervals from glm model
}
\details{
This is based on the \code{trending} package, as described here: https://cran.r-project.org/web/packages/trending/vignettes/prediction_intervals.html
}
