% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{check_columns}
\alias{check_columns}
\title{Check that dataframe contains appropriate columns}
\usage{
check_columns(input_data, column_names)
}
\arguments{
\item{input_data}{data.frame to check}

\item{column_names}{required column names}
}
\value{
names of missing columns (if missing), or 0 if none missing
}
\description{
Check that dataframe contains appropriate columns
}
