% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_ansi_log}
\alias{clean_ansi_log}
\title{Clean log output from R message formatting}
\usage{
clean_ansi_log(text_to_clean)
}
\arguments{
\item{text_to_clean}{string vector to clean}
}
\value{
cleaned text
}
\description{
Clean log output from R message formatting
}
