% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_forecast_quarto.R
\name{create_forecast_report}
\alias{create_forecast_report}
\title{Create HTML forecast report from quarto template}
\usage{
create_forecast_report(report_dir, quiet = TRUE)
}
\arguments{
\item{report_dir}{directory to build quarto doc in}

\item{quiet}{whether to render quarto doc quietly}
}
\value{
TRUE if successful, FALSE if error
}
\description{
Create HTML forecast report from quarto template
}
