% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_model_configs.R
\name{create_model_configs}
\alias{create_model_configs}
\title{Create default model configurations}
\usage{
create_model_configs(
  model_weights_df,
  W_graph,
  pred_vars,
  inla_hyper = list(prec.unstruct = c(1, 5e-04), prec.spatial = c(1, 5e-04), prec.timerw1
    = c(1, 0.01)),
  ranger_hyper = list(mtry = NULL, min.node.size = NULL, num.trees = 500)
)
}
\arguments{
\item{model_weights_df}{data.frame of model weights}

\item{W_graph}{graph used for INLA model}

\item{pred_vars}{formatted predictor variable names}

\item{inla_hyper}{list of inla hyper parameters. See Details}

\item{ranger_hyper}{list of ranger hyper parameters. See Details}
}
\value{
list of model configurations for each of five models
}
\description{
Create default model configurations
}
