% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_forecast_quarto.R
\name{create_simple_report}
\alias{create_simple_report}
\title{Simple report to provide when there is a quarto error}
\usage{
create_simple_report(report_dir)
}
\arguments{
\item{report_dir}{directory to save report in}
}
\value{
Nothing, but creates html document in \code{report_dir}
}
\description{
Simple report to provide when there is a quarto error
}
