% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_train_quarto.R
\name{create_train_quarto}
\alias{create_train_quarto}
\title{Function to create quarto doc of model training from model outputs}
\usage{
create_train_quarto(results_dir, html_filename, lang = "fr", doc_title = NULL)
}
\arguments{
\item{results_dir}{path to directory where model outputs are saved}

\item{html_filename}{where you want the html file to be saved, including path}

\item{doc_title}{title for HTML report. Default: "PRIDEC Rapport de Performance"}

\item{language}{which language template to use. currently only  "fr" available}
}
\description{
Function to create quarto doc of model training from model outputs
}
