% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_polygon}
\alias{demo_polygon}
\title{Demo malaria dataset polygon}
\format{
\subsection{\code{demo_polygon}}{

A simple feature geospatial collection to be used with \code{sf}. This is used by the examples of INLA models.
\describe{
\item{orgUnit}{DHIS2 orgUnit corresponding to primary health clinics}
\item{org_ID}{numeric id corresponding to orgUnit used internally by INLA}
\item{geom}{geometry of polygon associated with each orgUnit}
}
}
}
\source{
PRIDE-C research team
}
\usage{
demo_polygon
}
\description{
Polygons to used in spatial mesh by INLA in examples
}
\keyword{datasets}
