% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{eval_performance}
\alias{eval_performance}
\title{Evaluate performance on a cv_set}
\usage{
eval_performance(pred_intervals)
}
\arguments{
\item{pred_intervals}{data.frame of prediction intervals, output from \code{fit} step.
Must include: orgUnit, date, dataset, observed, predicted, quant_long, quantile}
}
\value{
dataset of performance metrics for assessment and anlalysis datasets for that cv_set
}
\description{
Evaluate performance on a cv_set
}
