% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{fill_seasonal}
\alias{fill_seasonal}
\title{Fill in missing data via seasonally decomposed imputation}
\usage{
fill_seasonal(data_to_fix, data_col, group_col, verbose = FALSE)
}
\arguments{
\item{data_to_fix}{dataframe of columns to impute, must contain date}

\item{data_col}{names of vector columns of data to fill}

\item{group_col}{name of column representing orgUnit grouping}

\item{verbose}{whether to print warnings. Default = FALSE}
}
\value{
original data_to_fix dataframe with missing data imputed
}
\description{
Fill in missing data via seasonally decomposed imputation
}
