% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arimax.R
\name{fit_arima}
\alias{fit_arima}
\title{Fit an ARIMAX model to a cv_set}
\usage{
fit_arima(
  cv_set,
  y_var,
  pred_vars,
  quant_levels = c(0.01, 0.025, seq(0.05, 0.95, by = 0.05), 0.975, 0.99)
)
}
\arguments{
\item{cv_set}{list containing \code{analysis} and \code{assessment} data.frames. One
element of the output of \code{\link[=split_cv_rolling]{split_cv_rolling()}}.}

\item{y_var}{character string of y variable label}

\item{pred_vars}{vector string of column names of predictor variables. Note
that static variables cannot be included in an ARIMAX model and will
be automatically removed.}

\item{quant_levels}{numeric vector of quantiles to use in prediction
intervals. Range 0-1. Default: c(0.01, 0.025, 0.05, 0.1, 0.15, 0.2, 0.25,
0.3, 0.35, 0.4, 0.45, 0.5, 0.55, 0.6, 0.65, 0.7, 0.75, 0.8, 0.85, 0.9,
0.95, 0.975, 0.99)}
}
\value{
data.frame of prediction intervals on analysis and assessment data
}
\description{
Fucntion to apply the ARIMAX workflow to each orgUnit in a dataset
}
