% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arimax.R
\name{fit_arima_OneOrgUnit}
\alias{fit_arima_OneOrgUnit}
\title{fit an ARIMAX model to one orgUnit}
\usage{
fit_arima_OneOrgUnit(
  train_df,
  test_df,
  pred_vars,
  quant_levels,
  return_model = FALSE
)
}
\arguments{
\item{train_df}{dataframe of analysis data}

\item{test_df}{dataframe of assessment data}

\item{pred_vars}{predictor variables to include}

\item{quant_levels}{quantiles needed for prediction intervals (assessment only)}

\item{return_model}{whether or not to return the ARIMA model, mostly used for debugging. Default = FALSE}
}
\value{
dataframe of prediction intervals for analysis and assessment data
}
\description{
fit an ARIMAX model to one orgUnit
}
