% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_nb.R
\name{fit_glm_nb}
\alias{fit_glm_nb}
\title{Fit a negative binomial GLM to a cv_set}
\usage{
fit_glm_nb(
  cv_set,
  y_var,
  id_vars,
  pred_vars,
  return_model = FALSE,
  quantile_levels = c(0.01, 0.025, seq(0.05, 0.95, by = 0.05), 0.975, 0.99)
)
}
\arguments{
\item{cv_set}{list object containing analysis and assessment data.frames}

\item{y_var}{string name of column of observed values}

\item{id_vars}{names of columns used as keys/ids in data.frames}

\item{pred_vars}{vector string of predictor variables to use in model}

\item{return_model}{whether or not to return ranger model object. Default  = FALSE.}

\item{quantile_levels}{quantile levels to use when predicted via quantile regression.
Default =  c(0.01,0.025, seq(0.05,0.95, by = 0.05), 0.975, 0.99)}
}
\value{
dataframe of prediction intervals
}
\description{
Fit a negative binomial GLM to a cv_set
}
