% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive.R
\name{fit_naive}
\alias{fit_naive}
\title{Fit a naive model to a cv_set}
\usage{
fit_naive(
  cv_set,
  y_var,
  group_vars = c("orgUnit", "month_season"),
  quantile_levels = c(0.01, 0.025, seq(0.05, 0.95, by = 0.05), 0.975, 0.99)
)
}
\arguments{
\item{cv_set}{list object containing analysis and assessment data.frames}

\item{y_var}{string name of column of observed values}

\item{group_vars}{string vector names of columns used to group means. Usually refers to month of year and orgUnit.}
}
\value{
data.frame of prediction intervals and observed values containing columns for group_vars and whether data is analysis or assessment
}
\description{
Fit a naive model to a cv_set
}
