% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{fit_ranger}
\alias{fit_ranger}
\title{Fit a ranger Random Forest model to a cv_set}
\usage{
fit_ranger(
  cv_set,
  y_var,
  id_vars,
  pred_vars,
  hyper_control = list(mtry = NULL, min.node.size = NULL, num.trees = 500),
  importance = "none",
  quantile_levels = c(0.01, 0.025, seq(0.05, 0.95, by = 0.05), 0.975, 0.99),
  return_model = FALSE
)
}
\arguments{
\item{cv_set}{list object containing analysis and assessment data.frames}

\item{y_var}{string name of column of observed values}

\item{id_vars}{names of columns used as keys/ids in data.frames}

\item{pred_vars}{vector string of predictor variables to use in model}

\item{hyper_control}{hyperparameters to use in the model, in either list or
named vector format. Defaults are ranger defaults}

\item{importance}{the method with which to estimate variable importance. Default = 'none' means
no variable importance is estimated}

\item{quantile_levels}{quantile levels to use when predicted via quantile regression.
Default =  c(0.01,0.025, seq(0.05,0.95, by = 0.05), 0.975, 0.99)}

\item{return_model}{whether or not to return ranger model object. Default  = FALSE.}
}
\value{
prediction intervals on analysis and assessment data or list containing model
object and prediction intervals, if \code{return_model} = TRUE
}
\description{
Fit a ranger Random Forest model to a cv_set
}
