% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_forecast_dhis.R
\name{format_forecast_dhis}
\alias{format_forecast_dhis}
\title{Format forecasts to be posted to DHIS2 instance}
\usage{
format_forecast_dhis(
  forecast_out,
  disease_dataElement,
  forecast_start,
  month_assess,
  month_analysis
)
}
\arguments{
\item{forecast_out}{data.frame output of \code{ensemble_forecast}}

\item{disease_dataElement}{dhis2 code that represents the dataElement to be forecast}

\item{forecast_start}{forecast start in period \code{YYYYMM} format}

\item{month_assess}{months of data to forecast forward. integer}

\item{month_analysis}{months of data to use to train the model. integer}
}
\value{
cleaned formatted data.frame for DHIS2
}
\description{
Format forecasts to be posted to DHIS2 instance
}
