% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pred_var_names.R
\name{format_pred_vars}
\alias{format_pred_vars}
\title{Format predictor variables names to account for lagging and scaling}
\usage{
format_pred_vars(var_list, input_data)
}
\arguments{
\item{var_list}{string vector of predictor variables}

\item{input_data}{data.frame of input data}
}
\value{
list of variable names, variables to lag, and variables to scale
}
\description{
Format predictor variables names to account for lagging and scaling
}
