% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arimax.R
\name{get_arima_pi}
\alias{get_arima_pi}
\title{Internal function for estimating prediction intervals from ARIMAX model}
\usage{
get_arima_pi(arima_mod, quant_levels, h = 3, xreg = NULL)
}
\arguments{
\item{arima_mod}{ARIMAX model object}

\item{quant_levels}{quantile levels to estimate for the prediction intervals. Numeric 0-1}

\item{h}{Arima forecast horizon, in months. Default = 3}

\item{xreg}{corresponds to exogeneous variables in ARIMAX model (optional)}
}
\value{
data.frame of prediction intervals with columns date, prediction,
quantile, and quant_long. Only for assessment data.
}
\description{
Internal function for estimating prediction intervals from ARIMAX model
}
