% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cv_subsets}
\alias{get_cv_subsets}
\title{Internal function to format cv_set for analysis}
\usage{
get_cv_subsets(cv_set, y_var, pred_vars = NULL, remove_NA = FALSE)
}
\arguments{
\item{cv_set}{list of analysis and assessment data.frames}

\item{y_var}{name of variable to predict}

\item{pred_vars}{predictor variables used in model, or grouping variables in the case of \code{\link[=fit_naive]{fit_naive()}}. Default = NULL}

\item{remove_NA}{whether to remove missing y observations from analysis data. Default = FALSE}
}
\value{
list of \code{analysis} and \code{assess} dataframes
}
\description{
Internal function to format cv_set for analysis
}
