% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla.R
\name{get_inla_pi}
\alias{get_inla_pi}
\title{Function to calculate and format posterior prediction distributions from INLA using random error trick}
\usage{
get_inla_pi(inla_marginal, quantile_levels, id_df)
}
\arguments{
\item{inla_marginal}{output of \code{marginals.fitted.values} from INLA model for one observation}

\item{quantile_levels}{quantiles that you want to estimate interval at}

\item{id_df}{a one row dataframe containing the id variables and true_value for that observation}
}
\description{
One way to calculate the PI for an INLA model with the appropriate error is to fit a random error
term to each observation. This takes the output of that to estimate PIs.
}
