% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla.R
\name{get_inla_pi_sample}
\alias{get_inla_pi_sample}
\title{Function to calculate and format posterior prediction distributions from INLA via posterior sampling}
\usage{
get_inla_pi_sample(inla_post_sample, quantile_levels, id_df)
}
\arguments{
\item{inla_post_sample}{samples for one predictor (often output of inla.posterior.sample.eval)}

\item{quantile_levels}{quantiles that you want to estimate interval at}

\item{id_df}{a one row dataframe containing the id variables and true_value for that observation}
}
\description{
Function to calculate and format posterior prediction distributions from INLA via posterior sampling
}
