% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{get_wis}
\alias{get_wis}
\title{Calculate Weighted Interval Score}
\usage{
get_wis(pred_df)
}
\arguments{
\item{pred_df}{data.frame of predictions included observed, predicted and quantile_level}
}
\description{
Internal function to use \code{wis} from \code{scoringutils}
}
