% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arimax.R
\name{inv_variables_arima}
\alias{inv_variables_arima}
\title{Investigate variables from ARIMAX model}
\usage{
inv_variables_arima(cv_set, y_var, pred_vars, var_scales)
}
\arguments{
\item{cv_set}{list containing \code{analysis} and \code{assessment} data.frames. One
element of the output of \code{\link[=split_cv_rolling]{split_cv_rolling()}}.}

\item{y_var}{character string of y variable label}

\item{pred_vars}{vector string of column names of predictor variables. Note
that static variables cannot be included in an ARIMAX model and will
be automatically removed.}

\item{var_scales}{data.frame of centering and scaling values for predictor variables}
}
\value{
list of variable importance and data for counter-factual plots
}
\description{
Investigate variables from ARIMAX model
}
