% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_nb.R
\name{inv_variables_glm_nb}
\alias{inv_variables_glm_nb}
\title{Estimate variable importance and counter-factual plots of a glm.nb model}
\usage{
inv_variables_glm_nb(
  cv_set,
  y_var,
  id_vars,
  pred_vars,
  var_scales,
  constant_org = NULL,
  constant_date = NULL,
  nsim = 50
)
}
\arguments{
\item{cv_set}{list object containing analysis and assessment data.frames}

\item{y_var}{string name of column of observed values}

\item{id_vars}{names of columns used as keys/ids in data.frames}

\item{pred_vars}{vector string of predictor variables to use in model}

\item{var_scales}{data.frame containing centering and scaling parameters for variables}

\item{constant_org}{which orgUnit to use for counterfactual plots. If NULL (default), chooses one
at random}

\item{constant_date}{whcih date to use for counterfactual plots. If NULL (default), is set to median of
analysis data.}

\item{nsim}{number of simulations to use in variable importance permutation. Default = 50.}
}
\value{
list containing variable importance scores and a list of dataframes
containing data for counterfactual plots with each element corresponding to a variable
}
\description{
Estimate variable importance and counter-factual plots of a glm.nb model
}
