% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla.R
\name{inv_variables_inla}
\alias{inv_variables_inla}
\title{Estimate variable importance and partial dependence plots of a  INLA model}
\usage{
inv_variables_inla(
  cv_set,
  y_var,
  pred_vars,
  reff_var = NULL,
  id_vars,
  hyper_priors = list(prec.unstruct = c(1, 5e-04), prec.spatial = c(1, 5e-04),
    prec.timerw1 = c(1, 0.01)),
  W_orgUnit,
  var_scales,
  constant_org,
  constant_date,
  seed = 8675309,
  nsims = 1
)
}
\arguments{
\item{cv_set}{list object containing analysis and assessment data.frames}

\item{y_var}{string name of column of observed values}

\item{pred_vars}{vector string of predictor variables to use in model}

\item{reff_var}{string or formula for random effects}

\item{id_vars}{names of columns used as keys/ids in data.frames}

\item{hyper_priors}{priors for spatial and temporal random effects}

\item{W_orgUnit}{graph of orgUnit to use in spatial structure. If NULL (default), will not fit a spatial structure.}

\item{var_scales}{data.frame containing centering and scaling parameters for variables}

\item{constant_org}{orgUnit to use in counterfactual plots}

\item{constant_date}{date to use in counterfactual plots}

\item{seed}{a seed to use for reproducability. Default = 8675309}

\item{nsims}{number of simulations to run. Default = 1}
}
\value{
list containing variable importance scores and a list of dataframes
containing data for pdp plots with each element corresponding to a variable
}
\description{
Estimate variable importance and partial dependence plots of a  INLA model
}
