% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{inv_variables_ranger}
\alias{inv_variables_ranger}
\title{Estimate variable importance and partial dependence plots of a ranger model}
\usage{
inv_variables_ranger(
  cv_set,
  y_var,
  id_vars,
  pred_vars,
  hyper_control = list(mtry = NULL, min.node.size = NULL, num.trees = 500),
  var_scales
)
}
\arguments{
\item{cv_set}{list object containing analysis and assessment data.frames}

\item{y_var}{string name of column of observed values}

\item{id_vars}{names of columns used as keys/ids in data.frames}

\item{pred_vars}{vector string of predictor variables to use in model}

\item{hyper_control}{hyperparameters to use in the model, in either list or
named vector format. Defaults are ranger defaults}

\item{var_scales}{data.frame containing centering and scaling parmaters for variables}
}
\value{
list containing variable importance scores and a list of dataframes
containing data for pdp plots with each element corresponding to a variable
}
\description{
Estimate variable importance and partial dependence plots of a ranger model
}
