% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_counterfactual}
\alias{plot_counterfactual}
\title{Plot multiple counterfactual plots from the list of cf_data using ggplot}
\usage{
plot_counterfactual(cf_list, y_range = NULL)
}
\arguments{
\item{cf_list}{list of counterfactual data.frames. Names correspond to variable labels}

\item{y_range}{range of y_variable to use to limit plots}
}
\value{
list of plots to be plotted with \code{patchwork::wrap_plots}
}
\description{
Plot multiple counterfactual plots from the list of cf_data using ggplot
}
