% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_counterfactual_one}
\alias{plot_counterfactual_one}
\title{Plot one counterfactual plot from models using ggplot}
\usage{
plot_counterfactual_one(cf_data, var_label, y_range = NULL)
}
\arguments{
\item{cf_data}{counterfactual data frame}

\item{var_label}{label used for predictor variable}

\item{y_range}{range of y_variable to use to limit plots}
}
\description{
Plot one counterfactual plot from models using ggplot
}
