% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_predictions}
\alias{plot_predictions}
\title{Plot predictions and PI ribbon with ggplot2}
\usage{
plot_predictions(pred_df, quantile_ribbon = c(0.05, 0.95))
}
\arguments{
\item{pred_df}{data.frame of prediction intervals}

\item{quantile_ribbon}{quantile levels to use for ribbon}
}
\value{
ggplot object
}
\description{
Plot predictions and PI ribbon with ggplot2
}
