% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Prepare case data and predictor variables for forecast models}
\usage{
prep_data(
  raw_data,
  y_var,
  lagged_vars,
  scaled_vars,
  lag_n = 3,
  graph_poly = NULL
)
}
\arguments{
\item{raw_data}{data.frame containing the variables}

\item{y_var}{name of variable to predict}

\item{lagged_vars}{string vector of names of variables to lag}

\item{scaled_vars}{string vector of names of variables to scale}

\item{lag_n}{number of months by which to lag the lagged variables. Default = 3.}

\item{graph_poly}{sf object to be used to create graph for INLA model
(optional). Default = NULL.}
}
\value{
list of cleaned data, spatial graph for INLA (if not null), and
scale and centering factors used in treatment
}
\description{
Prepare case data and predictor variables for forecast models
}
