% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pridec_forecast.R
\name{run_pridec_forecast}
\alias{run_pridec_forecast}
\title{Function to run a PRIDE-C forecast with validated inputs
Detail the steps here}
\usage{
run_pridec_forecast(inputs, output_dir)
}
\arguments{
\item{inputs}{list of config, input_data and orgUnit poly output from validate_inputs}

\item{output_dir}{directory to save outputs in. If NULL, saves in a temporary directory}
}
\value{
T for success, F for error
}
\description{
Function to run a PRIDE-C forecast with validated inputs
Detail the steps here
}
