% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla.R
\name{sample_fixed_posteriors}
\alias{sample_fixed_posteriors}
\title{Sample from fixed-effect posteriors}
\usage{
sample_fixed_posteriors(inla_model, n_samp = 10000)
}
\arguments{
\item{inla_model}{INLA model object}

\item{n_samp}{number of samples to take from posterior}
}
\description{
Sample from fixed-effect posteriors
}
