% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_forecast.R
\name{sample_preds}
\alias{sample_preds}
\title{Utility function to get weighted means of predictions}
\usage{
sample_preds(predicted, weight)
}
\arguments{
\item{predicted}{vector of predicted values}

\item{weight}{vector of weights}
}
\description{
Utility function to get weighted means of predictions
}
