% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_cv.R
\name{split_cv_rolling}
\alias{split_cv_rolling}
\title{Create cv folds within a dataset using rolling origin method}
\usage{
split_cv_rolling(
  data_to_split,
  start_date = NULL,
  month_analysis = 48,
  month_assess = 3
)
}
\arguments{
\item{data_to_split}{data frame that you want to create folds from, must include date column}

\item{start_date}{date from which the folds will start, in Date class. default NULL = first date of dataset}

\item{month_analysis}{number of months to include in analysis sets. Default = 48}

\item{month_assess}{number of months to include in the assessment sets. Default = 3}
}
\value{
list of datasets including analysis and assessment set nested in each list object
}
\description{
Create cv folds within a dataset using rolling origin method
}
