% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_cv.R
\name{split_stratified}
\alias{split_stratified}
\title{Split the dataset into a training and testing set}
\usage{
split_stratified(cvdata_list)
}
\arguments{
\item{cvdata_list}{list of cv folds (output of split_cv_rolling)}
}
\value{
list of cv folds to be used in training ("train_data") and testing ("test_data")
}
\description{
Split the dataset into a training and testing set
}
\details{
Dataset is stratified by year, with an assessment period beginning in four months
of every year
}
