% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_forest.R
\name{tune_ranger}
\alias{tune_ranger}
\title{Tune a ranger random forest model}
\usage{
tune_ranger(
  cvfold_list,
  metric = "wape",
  y_var,
  id_vars,
  pred_vars,
  tune_grid = NULL
)
}
\arguments{
\item{cvfold_list}{a nested list of cv_sets, each with analysis and assessment}

\item{metric}{metric to use to select best fit. Options are output of \code{eval_performance}.}

\item{y_var}{string name of column of observed values}

\item{id_vars}{names of columns used as keys/ids in data.frames}

\item{pred_vars}{vector string of predictor variables to use in model}

\item{tune_grid}{grid of parameters to tune over, with each column values for a parameter.
If Default (NULL) will create one within the function.}
}
\value{
vector of best fitting parameters
}
\description{
Tune a ranger random forest model
}
