% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_forecast_dhis.R
\name{validate_dhis_format}
\alias{validate_dhis_format}
\title{Validate formatted forecast data for DHIS2}
\usage{
validate_dhis_format(this_df, quiet = FALSE)
}
\arguments{
\item{this_df}{formatted DHIS2 forecast}

\item{quiet}{whether to print number of errors}
}
\value{
prints number of NA, non-integer, and negative values. These should all be zero
}
\description{
Validate formatted forecast data for DHIS2
}
