% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_inputs.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate input data}
\usage{
validate_inputs(
  config,
  external_data = NULL,
  disease_data,
  climate_data,
  orgUnit_poly,
  return_inputs = FALSE
)
}
\arguments{
\item{config}{list of configurations for forecase. See Details}

\item{external_data}{data.frame containing external data to use in prediction. Must contain columns orgUnit, period, and predictor variables specified in the config (optional)}

\item{disease_data}{data.frame containing historical disease data used to train model. Must contain columns orgUnit, period, dataElement, value}

\item{climate_data}{data.frame containing historical disease data used to train model. Must contain columns orgUnit, period, dataElement, value}

\item{orgUnit_poly}{sf data.frame containing geometry of orgUnits. Must contain column orgUnit}

\item{return_inputs}{whether to return validated input or not. If TRUE, returns list of configs, input data, and orgPoly}
}
\value{
T for success, F for error. Or if return_inputs=TRUE, list of validated inputs
}
\description{
Validates the input data for a PRIDE-C forecasting workflow
}
\details{
The \code{config} argument should have the following elements:
\itemize{
\item pred_vars (string): vector of predictor variable names
\item model_weights (data.frame): columns model and weight
\item inla_hyper (list) : list of inla hyperparameters
\item ranger_hyper (list) : list of ranger hyperparameters
\item quantile_levels (numeric, 3)
\item month_analysis (integer, 1)
\item month_assess (integer, 1)
\item month_lag (integer, 1). Defaults to 3
\item forecast_start (string): date in format YYYYMM
}
}
