% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla.R
\name{calc_inla_vi}
\alias{calc_inla_vi}
\title{Calculate variable importance in INLA via simulation}
\usage{
calc_inla_vi(
  cv_set,
  y_var,
  pred_vars,
  reff_var = NULL,
  id_vars,
  hyper_priors = list(prec.unstruct = c(1, 5e-04), prec.spatial = c(1, 5e-04),
    prec.timerw1 = c(1, 0.01)),
  W_orgUnit,
  seed = 8675309,
  nsims = 1
)
}
\arguments{
\item{cv_set}{list object containing analysis and assessment data.frames}

\item{y_var}{string name of column of observed values}

\item{pred_vars}{vector string of predictor variables to use in model}

\item{reff_var}{string or formula for random effects}

\item{id_vars}{names of columns used as keys/ids in data.frames}

\item{hyper_priors}{priors for spatial and temporal random effects}

\item{W_orgUnit}{graph of orgUnit to use in spatial structure}

\item{seed}{a seed to use for reproducability. Default = 8675309}

\item{nsims}{number of simulations to run. Default = 1}
}
\description{
Calculate variable importance in INLA via simulation
}
