% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla.R
\name{create_counterfactual_inla}
\alias{create_counterfactual_inla}
\title{Create counterfactual data of variables from INLA model}
\usage{
create_counterfactual_inla(
  cv_set,
  y_var,
  pred_vars,
  reff_var = NULL,
  id_vars,
  hyper_priors = list(prec.unstruct = c(1, 5e-04), prec.spatial = c(1, 5e-04),
    prec.timerw1 = c(1, 0.01)),
  W_orgUnit,
  var_scales,
  constant_org,
  constant_date
)
}
\arguments{
\item{cv_set}{list object containing analysis and assessment data.frames}

\item{y_var}{string name of column of observed values}

\item{pred_vars}{vector string of predictor variables to use in model}

\item{reff_var}{string or formula for random effects}

\item{id_vars}{names of columns used as keys/ids in data.frames}

\item{hyper_priors}{priors for spatial and temporal random effects}

\item{W_orgUnit}{graph of orgUnit to use in spatial structure. If NULL (default), will not fit a spatial structure.}

\item{var_scales}{data.frame of center and scale factors used in variable scaling}

\item{constant_org}{orgUnit to use in counterfactual plots}

\item{constant_date}{date to use in counterfactual plots}
}
\value{
list of dataframes of counter factual data. Each element corresponds to one predictor variable
}
\description{
Create counterfactual data of variables from INLA model
}
