% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{create_pridec_quarto}
\alias{create_pridec_quarto}
\title{Function to create quarto doc from model outputs}
\usage{
create_pridec_quarto(results_dir, html_filename)
}
\arguments{
\item{results_dir}{path to directory where model outputs are saved. cannot be relative}

\item{html_filename}{where you want the html file to be saved. cannot be relative}
}
\description{
Function to create quarto doc from model outputs
}
