% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_forecast.R
\name{ensemble_forecast}
\alias{ensemble_forecast}
\title{Fit an ensemble forecast}
\usage{
ensemble_forecast(
  cv_set,
  y_var,
  id_vars,
  quantile_levels = c(0.025, 0.5, 0.975),
  inla_configs = NULL,
  glm_nb_configs = NULL,
  ranger_configs = NULL,
  arimax_configs = NULL,
  naive_configs = NULL
)
}
\arguments{
\item{cv_set}{list of analysis and assessment data to forecast over}

\item{y_var}{name of y variable}

\item{id_vars}{name of columns used to identify each sample. Usually time period and orgUnit}

\item{quantile_levels}{quantile level of prediction intervals. Default = c(0.025, 0.5,0.975)
corresponding to 95\% CI}

\item{inla_configs}{named list of configuration for INLA model. Must include: reff_var, pred_vars,
hyper_priors, W_orgUnit, sample_pi. See \link[PRIDEC]{fit_inla} for more info.
If NULL (default), will not fit an inla model.}

\item{glm_nb_configs}{named list of configuration for GLM model. Must include: pred_vars}

\item{ranger_configs}{named list of configuation for ranger model. Must include: pred_vars,
hyper_control. See \link[PRIDEC]{fit_ranger} for more info.}

\item{arimax_configs}{named list of configuration for ARIMAX model. Must include: pred_vars,
log_trans. See \link[PRIDEC]{fit_arima} for more info.}

\item{naive_configs}{named list of configuration for naive model. Must include: group_vars.
See \link[PRIDEC]{fit_naive} for more info.}
}
\value{
Dataframe of prediction intervals for the ensembled predictions
}
\description{
Fit an ensemble forecast
}
