% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla.R
\name{fit_inla}
\alias{fit_inla}
\title{Fit an INLA model to one cv_set and get prediction intervals}
\usage{
fit_inla(
  cv_set,
  y_var,
  pred_vars,
  id_vars,
  reff_var = NULL,
  hyper_priors = list(prec.unstruct = c(1, 5e-04), prec.spatial = c(1, 5e-04),
    prec.timerw1 = c(1, 0.01)),
  return_model = FALSE,
  quantile_levels = c(0.01, 0.025, seq(0.05, 0.95, by = 0.05), 0.975, 0.99),
  sample_pi = FALSE,
  W_orgUnit = NULL,
  verbose = FALSE,
  n_cores = NULL
)
}
\arguments{
\item{cv_set}{list object containing analysis and assessment data.frames}

\item{y_var}{string name of column of observed values}

\item{pred_vars}{vector string of predictor variables to use in model}

\item{id_vars}{names of columns used as keys/ids in data.frames}

\item{reff_var}{string or formula for random effects}

\item{hyper_priors}{priors for spatial and temporal random effects}

\item{return_model}{whether or not to return the INLA model in addition to predictions}

\item{quantile_levels}{quantile levels to use when predicted via quantile regression.
Default =  c(0.01,0.025, seq(0.05,0.95, by = 0.05), 0.975, 0.99)}

\item{sample_pi}{whether to estimate the prediction intervals by sampling the
posterior. Default (FALSE) uses a trick to put all error into an iid random effect.}

\item{W_orgUnit}{graph of orgUnit to use in spatial structure. If NULL (default), will not fit a spatial structure.}

\item{verbose}{whether to run the INLA model verbosely. Default = FALSE}

\item{n_cores}{number of cores to use (integer). Defaults to number of available cores minus 2.}
}
\value{
dataframe of predictions intervals
}
\description{
Fit an INLA model to one cv_set and get prediction intervals
}
