% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIDEC-package.R
\docType{package}
\name{PRIDEC-package}
\alias{PRIDEC}
\alias{PRIDEC-package}
\title{PRIDEC: Disease Forecasting with PRIDE-C}
\description{
Functions used by the Predicting Infectious Diseases via Environment and Climate (PRIDE-C) forecasting workflow to forecast infectious diseases for DHIS2 instances.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pivot-madagascar.github.io/PRIDEC-package/}
}

}
\author{
\strong{Maintainer}: Michelle Evans \email{first.last@example.com} (\href{https://orcid.org/0000-0002-5628-0502}{ORCID})

}
\keyword{internal}
