% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla.R
\name{run_inla_config}
\alias{run_inla_config}
\title{Configuration for an INLA model
An internal function to run a configured INLA model}
\usage{
run_inla_config(
  formula,
  data,
  family = "zeroinflatednbinomial1",
  config = T,
  verbose = T
)
}
\arguments{
\item{formula}{forumula in y ~x format for model}

\item{data}{data to be used in model}

\item{family}{family of model to fit. Default = "zeroinflatednbinomial1"}

\item{config}{whetehr or not to use config for control.compute step. Default = TRUE}

\item{verbose}{whether fitting should be verbose. Default = TRUE}
}
\description{
Configuration for an INLA model
An internal function to run a configured INLA model
}
