% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_cv.R
\name{split_cv_forecast}
\alias{split_cv_forecast}
\title{Create data to use for forecast}
\usage{
split_cv_forecast(
  data_to_split,
  forecast_start_date,
  month_analysis,
  month_assess
)
}
\arguments{
\item{forecast_start_date}{start date of forecast}

\item{month_analysis}{how many months of analysis to include}

\item{month_assess}{how many months of forecasting (assessment to include)}

\item{all_data}{data.frame that contains the data used to train the model}
}
\value{
list containing analysis and assessment set
}
\description{
Create data to use for forecast
}
