% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_models.R
\name{train_models}
\alias{train_models}
\title{Run full train_model workflow using multiple modeling approaches}
\usage{
train_models(
  prep_output,
  models,
  y_var,
  pred_vars,
  id_vars = c("orgUnit", "date"),
  results_dir = NULL,
  tune = NULL,
  model_configs
)
}
\arguments{
\item{prep_output}{output of prep_data}

\item{models}{vector of models to fit. Options: naive, arimax, glm_nb, ranger, inla}

\item{y_var}{character. name of variable to predict}

\item{pred_vars}{character vector of predictor variables. Not all will be used
in all models}

\item{results_dir}{directory to save model results and quarto document in.
If NULL, temporary directory will be used}

\item{tune}{string vector containing names of models to tune. Options = ranger}

\item{model_configs}{Optional list of configurations for each model. Name of
element in list should match model name. See \link[PRIDEC]{ensemble_forecast}
for more info.}
}
\value{
saves model outputs to \code{results_dir} if specified. Creates a quarto doc of
model outputs
}
\description{
Run full train_model workflow using multiple modeling approaches
}
